Update def/use calculation. Bug when calculating lower bounds.

+++ Generic/llvmTip/llvm-master/include/llvm/CodeGen/OptSched/generic/defines.h
@@ -22,7 +22,7 @@ namespace opt_sched {
 typedef int64_t Milliseconds;
 
 // Instruction count.
-typedef int64_t InstCount;
+typedef int InstCount;
 
 // A generic sentinel value. Should be used with care.
 // TODO(max): Get rid of this in favor of type- or purpose-specific sentinels.
